[test] gathering blindsign tests
